fix: renamed components
